% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alexnet_train.R
\name{alexnet_train}
\alias{alexnet_train}
\title{AlexNet Model}
\usage{
alexnet_train(
  batch_size = 32L,
  epochs = 2,
  data = NULL,
  strategy = NULL,
  parallel = NULL
)
}
\description{
Make use of AlexNet \link{https://papers.nips.cc/paper/4824-imagenet-classification-with-deep-convolutional-neural-networks.pdf}
to train an image classifier.
}
