% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alexnet_partition.R
\name{imagenet_partition}
\alias{imagenet_partition}
\title{Retrieve ImageNet Partition}
\usage{
imagenet_partition(
  partition,
  total = 16,
  url = "https://storage.googleapis.com/r-imagenet/"
)
}
\description{
Retrieves a partition to be used with \code{imagenet_train} as the \code{data}
parameter.
}
